+++ R/utils.R
@@ -156,3 +156,20 @@ compact <- function(x) {
   is_null <- vapply(x, is.null, logical(1))
   x[!is_null]
 }
+
+as_map <- function(x) {
+  # Remove nulls
+  is_null <- vapply(x, is.null, logical(1))
+  x <- x[!is_null]
+
+  # Sort named components, preserving positions of unnamed
+  nx <- names2(x)
+  is_named <- nx != ""
+  if (any(is_named)) {
+    idx <- seq_along(x)
+    idx[is_named] <- idx[is_named][order(nx[is_named])]
+    x <- x[idx]
+  }
+
+  x
+}